+++ packages/cli/src/controller/publish-controller.ts
@@ -3,7 +3,13 @@
 
 import fs from 'fs';
 import path from 'path';
-import {ReaderFactory, IPFS_WRITE_ENDPOINT, isFileReference, validateCommonProjectManifest, mapToObject} from '@subql/common';
+import {
+  ReaderFactory,
+  IPFS_WRITE_ENDPOINT,
+  isFileReference,
+  validateCommonProjectManifest,
+  mapToObject,
+} from '@subql/common';
 import {parseAlgorandProjectManifest} from '@subql/common-algorand';
 import {parseCosmosProjectManifest} from '@subql/common-cosmos';
 import {parseEthereumProjectManifest} from '@subql/common-ethereum';