+++ packages/node/CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.1.0] - 2023-10-20
+### Added
+- Inject in-memory cache to sandbox (#2110)
+### Fixed
+- Bump with `@subq/node-core` 3.1.0 , fixed poi migration init check, and improve logging
+
+## [3.0.8] - 2023-10-18
+### Fixed
+- Update node-core, fix store bulk methods failing with workers
+
 ## [3.0.7] - 2023-10-17
 ### Changed
 - Update with node-core 6.0.3
@@ -1011,7 +1021,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - bump @polkadot/api to 3.1.1
 
-[Unreleased]: https://github.com/subquery/subql/compare/node/3.0.7...HEAD
+[Unreleased]: https://github.com/subquery/subql/compare/v3.1.0...HEAD
+[3.1.0]: https://github.com/subquery/subql/compare/v3.0.8...v3.1.0
+[3.0.8]: https://github.com/subquery/subql/compare/node/3.0.7...node/3.0.8
 [3.0.7]: https://github.com/subquery/subql/compare/node/3.0.6...node/3.0.7
 [3.0.6]: https://github.com/subquery/subql/compare/node/3.0.5...node/3.0.6
 [3.0.5]: https://github.com/subquery/subql/compare/node/3.0.4...node/3.0.5